+++ CHANGELOG.md
@@ -99,7 +99,9 @@
 - Bugfix when parcel terminal unconditionally changed to first selection while searching.
 - Upgrade added to avoid error when client upgrades modules and does not delete cache.
 - Bugfix when parcel tracking validation was called from prestashop core. Fall back to universal legacy function
+- Logic to execute code further if error occurs on tracking email send.
 - Bugfix when parcel terminal form is always visible in order
 - Translatable string added when no option available, translations updated for all baltic languages for pickup select.
+